refactor code to improve stdout
