--- ikctl/local/run_kits.py
@@ -1,2 +1,0 @@
- # Run kits in local servers
-
